requirements.txt
requests
python-dotenv
langchain-core
langchain-upstage